repfun: Create TLFs using Functions Styled after SAS™ Macros

Mimic the style of traditional reporting macros for clinical trials. The purpose is to generate TLFs (tables, listings and figures) that support clinical research. This package is well suited for firms or individuals who wish to incorporate R without changing their ways of working as it follows a traditional clinical research workflow. Invoke functions (instead of macros) to summarize data and produce formatted reports. This package differs from others in that it includes tools (wrappers) for both analyzing and reporting data.

See also

Author

Maintainer: Chris Rook cr883296@gmail.com

Authors:

Yongwei Wang yongwei.x.wang@viivhealthcare.com

Other contributors:

GlaxoSmithKline Research & Development Limited [copyright holder, funder]