Generate proc contents of data frame similar to that generated by SAS

Pass in a data frame and the contents will be displayed.

Usage

ru_contents(dsetin)

Arguments

dsetin

Incoming data frame to have proc contents generated.

Value

The proc contents output will be displayed.

Author

Yongwei Wang, yongwei.x.wang@viivhealthcare.com
Chris Rook, cr883296@gmail.com

Examples

library(repfun)
repfun::ru_contents(mtcars)
#> [1] "MADE IT HERE ..."
#> 
#>  Data Set Name       mtcars                                            
#>  Member Type         Data                                              
#>  Created             NA                                                
#>  Last Modified       NA                                                
#>  Label                                                                 
#>  Observations        32                                                
#>  Variables           11                                                
#>  Filename            mtcars                                            
#>  File Size           0 MB                                              
#>  File Size (bytes)   7208                                              
#> 
#> 
#>      Variable  Type       Len Class       Format      Label                                             
#> 
#>    9 am        double       9                                                                           
#>   11 carb      double       9                                                                           
#>    2 cyl       double       9                                                                           
#>    3 disp      double       9                                                                           
#>    5 drat      double       9                                                                           
#>   10 gear      double       9                                                                           
#>    4 hp        double       9                                                                           
#>    1 mpg       double       9                                                                           
#>    7 qsec      double       9                                                                           
#>    8 vs        double       9                                                                           
#>    6 wt        double       9