Append data sets even when variables do not match

Pass in a collection of data frames separated by commas and they will be appended.

Usage

ru_setdata(..., keeprownames = TRUE)

Arguments

...

A collection of data frames.

keeprownames

Convert row names on data frame to columns and keep the column.

Value

The incoming data frames combined.

Author

Yongwei Wang, yongwei.x.wang@viivhealthcare.com
Chris Rook, christopher.x.rook@gsk.com

Examples

library(repfun)
repfun::ru_setdata(mtcars,airquality)
#>               Row.names  mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1             Mazda RX4 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2         Mazda RX4 Wag 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3            Datsun 710 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4        Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5     Hornet Sportabout 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6               Valiant 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7            Duster 360 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8             Merc 240D 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 9              Merc 230 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10             Merc 280 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 11            Merc 280C 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 12           Merc 450SE 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 13           Merc 450SL 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 14          Merc 450SLC 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15   Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 16  Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 17    Chrysler Imperial 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 18             Fiat 128 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 19          Honda Civic 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 20       Toyota Corolla 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 21        Toyota Corona 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 22     Dodge Challenger 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 23          AMC Javelin 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24           Camaro Z28 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 25     Pontiac Firebird 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 26            Fiat X1-9 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 27        Porsche 914-2 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 28         Lotus Europa 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 29       Ford Pantera L 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 30         Ferrari Dino 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 31        Maserati Bora 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 32           Volvo 142E 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 33                    1   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 34                    2   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 35                    3   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 36                    4   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 37                    5   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 38                    6   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 39                    7   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 40                    8   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 41                    9   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 42                   10   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 43                   11   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 44                   12   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 45                   13   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 46                   14   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 47                   15   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 48                   16   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 49                   17   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 50                   18   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 51                   19   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 52                   20   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 53                   21   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 54                   22   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 55                   23   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 56                   24   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 57                   25   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 58                   26   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 59                   27   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 60                   28   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 61                   29   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 62                   30   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 63                   31   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 64                   32   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 65                   33   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 66                   34   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 67                   35   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 68                   36   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 69                   37   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 70                   38   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 71                   39   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 72                   40   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 73                   41   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 74                   42   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 75                   43   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 76                   44   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 77                   45   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 78                   46   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 79                   47   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 80                   48   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 81                   49   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 82                   50   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 83                   51   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 84                   52   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 85                   53   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 86                   54   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 87                   55   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 88                   56   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 89                   57   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 90                   58   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 91                   59   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 92                   60   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 93                   61   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 94                   62   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 95                   63   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 96                   64   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 97                   65   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 98                   66   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 99                   67   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 100                  68   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 101                  69   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 102                  70   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 103                  71   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 104                  72   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 105                  73   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 106                  74   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 107                  75   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 108                  76   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 109                  77   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 110                  78   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 111                  79   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 112                  80   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 113                  81   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 114                  82   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 115                  83   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 116                  84   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 117                  85   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 118                  86   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 119                  87   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 120                  88   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 121                  89   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 122                  90   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 123                  91   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 124                  92   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 125                  93   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 126                  94   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 127                  95   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 128                  96   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 129                  97   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 130                  98   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 131                  99   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 132                 100   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 133                 101   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 134                 102   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 135                 103   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 136                 104   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 137                 105   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 138                 106   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 139                 107   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 140                 108   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 141                 109   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 142                 110   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 143                 111   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 144                 112   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 145                 113   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 146                 114   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 147                 115   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 148                 116   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 149                 117   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 150                 118   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 151                 119   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 152                 120   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 153                 121   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 154                 122   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 155                 123   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 156                 124   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 157                 125   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 158                 126   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 159                 127   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 160                 128   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 161                 129   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 162                 130   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 163                 131   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 164                 132   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 165                 133   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 166                 134   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 167                 135   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 168                 136   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 169                 137   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 170                 138   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 171                 139   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 172                 140   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 173                 141   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 174                 142   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 175                 143   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 176                 144   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 177                 145   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 178                 146   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 179                 147   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 180                 148   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 181                 149   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 182                 150   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 183                 151   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 184                 152   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#> 185                 153   NA  NA    NA  NA   NA    NA    NA NA NA   NA   NA
#>     Ozone Solar.R Wind Temp Month Day
#> 1      NA      NA   NA   NA    NA  NA
#> 2      NA      NA   NA   NA    NA  NA
#> 3      NA      NA   NA   NA    NA  NA
#> 4      NA      NA   NA   NA    NA  NA
#> 5      NA      NA   NA   NA    NA  NA
#> 6      NA      NA   NA   NA    NA  NA
#> 7      NA      NA   NA   NA    NA  NA
#> 8      NA      NA   NA   NA    NA  NA
#> 9      NA      NA   NA   NA    NA  NA
#> 10     NA      NA   NA   NA    NA  NA
#> 11     NA      NA   NA   NA    NA  NA
#> 12     NA      NA   NA   NA    NA  NA
#> 13     NA      NA   NA   NA    NA  NA
#> 14     NA      NA   NA   NA    NA  NA
#> 15     NA      NA   NA   NA    NA  NA
#> 16     NA      NA   NA   NA    NA  NA
#> 17     NA      NA   NA   NA    NA  NA
#> 18     NA      NA   NA   NA    NA  NA
#> 19     NA      NA   NA   NA    NA  NA
#> 20     NA      NA   NA   NA    NA  NA
#> 21     NA      NA   NA   NA    NA  NA
#> 22     NA      NA   NA   NA    NA  NA
#> 23     NA      NA   NA   NA    NA  NA
#> 24     NA      NA   NA   NA    NA  NA
#> 25     NA      NA   NA   NA    NA  NA
#> 26     NA      NA   NA   NA    NA  NA
#> 27     NA      NA   NA   NA    NA  NA
#> 28     NA      NA   NA   NA    NA  NA
#> 29     NA      NA   NA   NA    NA  NA
#> 30     NA      NA   NA   NA    NA  NA
#> 31     NA      NA   NA   NA    NA  NA
#> 32     NA      NA   NA   NA    NA  NA
#> 33     41     190  7.4   67     5   1
#> 34     36     118  8.0   72     5   2
#> 35     12     149 12.6   74     5   3
#> 36     18     313 11.5   62     5   4
#> 37     NA      NA 14.3   56     5   5
#> 38     28      NA 14.9   66     5   6
#> 39     23     299  8.6   65     5   7
#> 40     19      99 13.8   59     5   8
#> 41      8      19 20.1   61     5   9
#> 42     NA     194  8.6   69     5  10
#> 43      7      NA  6.9   74     5  11
#> 44     16     256  9.7   69     5  12
#> 45     11     290  9.2   66     5  13
#> 46     14     274 10.9   68     5  14
#> 47     18      65 13.2   58     5  15
#> 48     14     334 11.5   64     5  16
#> 49     34     307 12.0   66     5  17
#> 50      6      78 18.4   57     5  18
#> 51     30     322 11.5   68     5  19
#> 52     11      44  9.7   62     5  20
#> 53      1       8  9.7   59     5  21
#> 54     11     320 16.6   73     5  22
#> 55      4      25  9.7   61     5  23
#> 56     32      92 12.0   61     5  24
#> 57     NA      66 16.6   57     5  25
#> 58     NA     266 14.9   58     5  26
#> 59     NA      NA  8.0   57     5  27
#> 60     23      13 12.0   67     5  28
#> 61     45     252 14.9   81     5  29
#> 62    115     223  5.7   79     5  30
#> 63     37     279  7.4   76     5  31
#> 64     NA     286  8.6   78     6   1
#> 65     NA     287  9.7   74     6   2
#> 66     NA     242 16.1   67     6   3
#> 67     NA     186  9.2   84     6   4
#> 68     NA     220  8.6   85     6   5
#> 69     NA     264 14.3   79     6   6
#> 70     29     127  9.7   82     6   7
#> 71     NA     273  6.9   87     6   8
#> 72     71     291 13.8   90     6   9
#> 73     39     323 11.5   87     6  10
#> 74     NA     259 10.9   93     6  11
#> 75     NA     250  9.2   92     6  12
#> 76     23     148  8.0   82     6  13
#> 77     NA     332 13.8   80     6  14
#> 78     NA     322 11.5   79     6  15
#> 79     21     191 14.9   77     6  16
#> 80     37     284 20.7   72     6  17
#> 81     20      37  9.2   65     6  18
#> 82     12     120 11.5   73     6  19
#> 83     13     137 10.3   76     6  20
#> 84     NA     150  6.3   77     6  21
#> 85     NA      59  1.7   76     6  22
#> 86     NA      91  4.6   76     6  23
#> 87     NA     250  6.3   76     6  24
#> 88     NA     135  8.0   75     6  25
#> 89     NA     127  8.0   78     6  26
#> 90     NA      47 10.3   73     6  27
#> 91     NA      98 11.5   80     6  28
#> 92     NA      31 14.9   77     6  29
#> 93     NA     138  8.0   83     6  30
#> 94    135     269  4.1   84     7   1
#> 95     49     248  9.2   85     7   2
#> 96     32     236  9.2   81     7   3
#> 97     NA     101 10.9   84     7   4
#> 98     64     175  4.6   83     7   5
#> 99     40     314 10.9   83     7   6
#> 100    77     276  5.1   88     7   7
#> 101    97     267  6.3   92     7   8
#> 102    97     272  5.7   92     7   9
#> 103    85     175  7.4   89     7  10
#> 104    NA     139  8.6   82     7  11
#> 105    10     264 14.3   73     7  12
#> 106    27     175 14.9   81     7  13
#> 107    NA     291 14.9   91     7  14
#> 108     7      48 14.3   80     7  15
#> 109    48     260  6.9   81     7  16
#> 110    35     274 10.3   82     7  17
#> 111    61     285  6.3   84     7  18
#> 112    79     187  5.1   87     7  19
#> 113    63     220 11.5   85     7  20
#> 114    16       7  6.9   74     7  21
#> 115    NA     258  9.7   81     7  22
#> 116    NA     295 11.5   82     7  23
#> 117    80     294  8.6   86     7  24
#> 118   108     223  8.0   85     7  25
#> 119    20      81  8.6   82     7  26
#> 120    52      82 12.0   86     7  27
#> 121    82     213  7.4   88     7  28
#> 122    50     275  7.4   86     7  29
#> 123    64     253  7.4   83     7  30
#> 124    59     254  9.2   81     7  31
#> 125    39      83  6.9   81     8   1
#> 126     9      24 13.8   81     8   2
#> 127    16      77  7.4   82     8   3
#> 128    78      NA  6.9   86     8   4
#> 129    35      NA  7.4   85     8   5
#> 130    66      NA  4.6   87     8   6
#> 131   122     255  4.0   89     8   7
#> 132    89     229 10.3   90     8   8
#> 133   110     207  8.0   90     8   9
#> 134    NA     222  8.6   92     8  10
#> 135    NA     137 11.5   86     8  11
#> 136    44     192 11.5   86     8  12
#> 137    28     273 11.5   82     8  13
#> 138    65     157  9.7   80     8  14
#> 139    NA      64 11.5   79     8  15
#> 140    22      71 10.3   77     8  16
#> 141    59      51  6.3   79     8  17
#> 142    23     115  7.4   76     8  18
#> 143    31     244 10.9   78     8  19
#> 144    44     190 10.3   78     8  20
#> 145    21     259 15.5   77     8  21
#> 146     9      36 14.3   72     8  22
#> 147    NA     255 12.6   75     8  23
#> 148    45     212  9.7   79     8  24
#> 149   168     238  3.4   81     8  25
#> 150    73     215  8.0   86     8  26
#> 151    NA     153  5.7   88     8  27
#> 152    76     203  9.7   97     8  28
#> 153   118     225  2.3   94     8  29
#> 154    84     237  6.3   96     8  30
#> 155    85     188  6.3   94     8  31
#> 156    96     167  6.9   91     9   1
#> 157    78     197  5.1   92     9   2
#> 158    73     183  2.8   93     9   3
#> 159    91     189  4.6   93     9   4
#> 160    47      95  7.4   87     9   5
#> 161    32      92 15.5   84     9   6
#> 162    20     252 10.9   80     9   7
#> 163    23     220 10.3   78     9   8
#> 164    21     230 10.9   75     9   9
#> 165    24     259  9.7   73     9  10
#> 166    44     236 14.9   81     9  11
#> 167    21     259 15.5   76     9  12
#> 168    28     238  6.3   77     9  13
#> 169     9      24 10.9   71     9  14
#> 170    13     112 11.5   71     9  15
#> 171    46     237  6.9   78     9  16
#> 172    18     224 13.8   67     9  17
#> 173    13      27 10.3   76     9  18
#> 174    24     238 10.3   68     9  19
#> 175    16     201  8.0   82     9  20
#> 176    13     238 12.6   64     9  21
#> 177    23      14  9.2   71     9  22
#> 178    36     139 10.3   81     9  23
#> 179     7      49 10.3   69     9  24
#> 180    14      20 16.6   63     9  25
#> 181    30     193  6.9   70     9  26
#> 182    NA     145 13.2   77     9  27
#> 183    14     191 14.3   75     9  28
#> 184    18     131  8.0   76     9  29
#> 185    20     223 11.5   68     9  30